working on comp python 3

+++ Source/PyBoy/GameWindow/FrameBuffer/framebuffer.py
@@ -17,7 +17,7 @@ class FrameBuffer(object):
         self._array = array
         self._scaleFactor = scaleFactor
 
-        self._cache = np.zeros(tuple(x / scaleFactor for x in array.shape), dtype=np.uint32)
+        self._cache = np.zeros(tuple(int(x / scaleFactor) for x in array.shape), dtype=np.uint32)
 
     def fill(self, val):
         self._cache.fill(val)
